requirements.txt
python-dotenv
pydantic
requests
uvicorn
fastapi
fastapi-cors
cryptography
telethon
openai
pillow